+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Release date: TBD
 
 * Dropped support of end of life MySQL version 5.6
-* Dropped support of end of life MariaDB versions below 10.2
+* Dropped support of end of life MariaDB versions below 10.3
 * Dropped support of end of life Python version 3.6